+++ spans/__init__.py
@@ -12,7 +12,7 @@ together.
 """
 
 
-__version__ = "0.5.0"
+__version__ = "1.0.0"
 
 
 __all__ = [